requirements.txt
naff>=2.0.0
python-dotenv>=0.20.0
orjson>=3.6.7